プレゼンテーション層(コントローラ)の実装
