+++ src/dataaccess/connection.ts
@@ -1,0 +1,10 @@
+import mysql from "mysql2/promise";
+
+export async function connectMySQL() {
+  return await mysql.createConnection({
+    host: "localhost",
+    database: "reversi",
+    user: "reversi",
+    password: "password",
+  });
+}
