+++ src/main.ts
@@ -1,27 +1,8 @@
 import express from "express";
 import morgan from "morgan";
 import "express-async-errors";
-import mysql from "mysql2/promise";
-import { createConnection } from "mysql2";
-import { GameGateway } from "./dataaccess/gameGateway";
-import { TurnGateway } from "./dataaccess/turnGateway";
-import { MoveGateway } from "./dataaccess/moveGateway";
-import { SquareGateway } from "./dataaccess/squareGateway";
-
-const EMPTY = 0;
-const DARK = 1;
-const LIGHT = 2;
-
-const INITIAL_BOARD = [
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, DARK, LIGHT, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, LIGHT, DARK, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-  [EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY, EMPTY],
-];
+import { gameRouter } from "./presentation/gameRouter";
+import { turnRouter } from "./presentation/turnRouter";
 
 const PORT = 3000;
 const app = express();
@@ -29,138 +10,8 @@ const app = express();
 app.use(morgan("dev"));
 app.use(express.static("static", { extensions: ["html"] }));
 app.use(express.json());
-
-const gameGateway = new GameGateway();
-const turnGateway = new TurnGateway();
-const moveGateway = new MoveGateway();
-const squareGateway = new SquareGateway();
-
-app.post("/api/games", async (req, res) => {
-  const now = new Date();
-  const conn = await connectMySQL();
-  try {
-    await conn.beginTransaction();
-    const gameRecord = await gameGateway.insert(conn, now);
-    const turnRecord = await turnGateway.insert(
-      conn,
-      gameRecord.id,
-      0,
-      DARK,
-      now
-    );
-    const squareCount = INITIAL_BOARD.map((line) => line.length).reduce(
-      (v1, v2) => v1 + v2,
-      0
-    );
-    await squareGateway.insertAll(conn, turnRecord.id, INITIAL_BOARD);
-
-    await conn.commit();
-  } finally {
-    await conn.end();
-  }
-
-  res.status(201).end();
-});
-
-app.get("/api/games/latest/turns/:turnCount", async (req, res) => {
-  const turnCount = parseInt(req.params.turnCount);
-  const conn = await connectMySQL();
-  try {
-    const gameRecord = await gameGateway.findLatest(conn);
-    if (!gameRecord) {
-      throw new Error("Game not found");
-    }
-
-    const turnRecord = await turnGateway.findForGameIdAndTurnCount(
-      conn,
-      gameRecord.id,
-      turnCount
-    );
-    if (!turnRecord) {
-      throw new Error("Turn not found");
-    }
-
-    const squareRecords = await squareGateway.findForTurnId(
-      conn,
-      turnRecord.id
-    );
-
-    const board = Array.from(Array(8)).map(() => Array.from(Array(8)));
-    squareRecords.forEach((s) => {
-      board[s.y][s.x] = s.disc;
-    });
-    const responseBody = {
-      turnCount: turnCount,
-      board: board,
-      nextDisc: turnRecord.nextDisc,
-      winnerDisc: null,
-    };
-    res.json(responseBody);
-  } finally {
-    await conn.end();
-  }
-});
-
-app.post("/api/games/latest/turns", async (req, res) => {
-  const turnCount = parseInt(req.body.turnCount);
-  const disc = parseInt(req.body.move.disc);
-  const x = parseInt(req.body.move.x);
-  const y = parseInt(req.body.move.y);
-
-  //dbから前のターンの情報を取得
-  const conn = await connectMySQL();
-  try {
-    const gameRecord = await gameGateway.findLatest(conn);
-    if (!gameRecord) {
-      throw new Error("Game not found");
-    }
-
-    const previousturnCount = turnCount - 1;
-    let previousTurnRecord = await turnGateway.findForGameIdAndTurnCount(
-      conn,
-      gameRecord.id,
-      previousturnCount
-    );
-    if (!previousTurnRecord) {
-      throw new Error("Turn not found");
-    }
-
-    const squareRecords = await squareGateway.findForTurnId(
-      conn,
-      previousTurnRecord.id
-    );
-
-    const board = Array.from(Array(8)).map(() => Array.from(Array(8)));
-    squareRecords.forEach((s) => {
-      board[s.y][s.x] = s.disc;
-    });
-
-    //石が置けるかどうか
-
-    //石をおく
-    board[y][x] = disc;
-    // dbに保存する
-    const now = new Date();
-
-    const nextDisc = disc === DARK ? LIGHT : DARK;
-    const turnRecord = await turnGateway.insert(
-      conn,
-      gameRecord.id,
-      turnCount,
-      nextDisc,
-      now
-    );
-
-    await squareGateway.insertAll(conn, turnRecord.id, board);
-    await moveGateway.insert(conn, turnRecord.id, disc, x, y);
-    await conn.commit();
-
-    res.status(201).end();
-  } finally {
-    await conn.end();
-  }
-});
-
+app.use(gameRouter);
+app.use(turnRouter);
 app.use(errorHandler);
 
 app.listen(PORT, () => {
@@ -176,14 +27,5 @@ function errorHandler(
   console.error("Unexpected error occurred", err);
   res.status(500).send({
     message: "Unexpected error occurred",
-  });
-}
-
-async function connectMySQL() {
-  return await mysql.createConnection({
-    host: "localhost",
-    database: "reversi",
-    user: "reversi",
-    password: "password",
   });
 }
